Close #5440 - java.io.FileNotFoundException in testPlatform for new MB project
 * check the folder exists before the copy

+++ Studio/src/com/twinsoft/convertigo/engine/admin/services/mobiles/MobileResourceHelper.java
@@ -114,7 +114,9 @@ public class MobileResourceHelper {
 			FileUtils.deleteQuietly(destDir);
 			for (File directory: mobileDir) {
 				currentMobileDir = directory;
-				FileUtils.copyDirectory(directory, destDir, fileFilterForCopy, true);
+				if (directory.exists()) {
+					FileUtils.copyDirectory(directory, destDir, fileFilterForCopy, true);
+				}
 			}
 			currentMobileDir = null;
